requirements.txt
fastapi
uvicorn
pydantic
pymysql
transformers
google-generativeai